pyproject.toml
[project]
name = "uiclient"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.4",
    "groq>=0.11.0",
    "openai>=1.54.3",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.17",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "uvicorn>=0.32.0",
]
